Add NotificationManager
